remove gulp, npm scripts
