+++ package.json
@@ -8,8 +8,14 @@
     "url": "git://github.com/menpo/landmarker.git"
   },
   "scripts": {
-    "build": "NODE_ENV=production gulp build",
-    "watch": "gulp",
+    "build:clean": "rm -rf ./build",
+    "build:copystatic": "cp -r ./static ./build",
+    "build:prep": "npm run build:clean && npm run build:copystatic",
+    "build:webpack": "webpack --config ./webpack/webpack.prod.config.js",
+    "build:appcache": "sed -i '' 's/<html lang=\"en\">/<html lang=\"en\" manifest=\"lmio.appcache\">/g' ./build/index.html",
+    "build": "NODE_ENV=production npm run build:prep && npm run build:webpack && npm run build:appcache",
+    "watch:webpack-dev-server": "webpack-dev-server --config ./webpack/webpack.dev.config.js --colours --content-base ./build",
+    "watch": "npm run build:prep && npm run watch:webpack-dev-server",
     "test": "mocha --compilers js:babel/register -R spec --recursive ./test",
     "coverage": "istanbul cover _mocha -- --compilers js:babel/register -R dot --recursive ./test",
     "lint": "eslint -c .eslintrc src/js",
@@ -27,14 +33,9 @@
     "autoprefixer-loader": "^3.1.0",
     "babel": "^5.6.3",
     "babel-loader": "^5.3.2",
-    "babelify": "^6.1.2",
     "backbone": "^1.1.2",
     "css-loader": "^0.21.0",
-    "del": "^1.1.1",
     "file-loader": "^0.8.4",
-    "gulp": "^3.8.10",
-    "gulp-replace": "^0.5.4",
-    "gulp-util": "^3.0.2",
     "jquery": "^2.1.3",
     "js-yaml": "^3.3.1",
     "json-loader": "^0.5.3",
