+++ webpack/webpack.dev.config.js
@@ -1,0 +1,5 @@
+var webpackConfig = require("./webpack.base.config.js");
+webpackConfig.devtool = "eval-source-map";
+webpackConfig.debug = true;
+
+module.exports = webpackConfig;
